Don't allow resize() on anonymous memory
This seems like an unfixed bug in Python 3.5. ftruncate(2) is called on
file descriptor -1, which results in an OSError. This seems to hard to
fix with the current approach of subclassing the built-in mmap class. A
complete reimplementation will be required so that we can implement
mremap() and keep the returned pointer.