Fix repr() test on Python 3.9
The output uses the internal class name. For now I prefer to play safe
and only override the built-in implementation when I have a good reason,
so I defer to the built-in repr() for 3.9. That requires this change so
that the test can work with both overriden and built-in repr().