Fallback implementation for platforms without memrchr(3)
This has a #define which is used to determine if we should fall back to
a local implementation of memrchr(3). Since some platforms do not have
memrchr(3) at all, we provide an empty implementation just to keep the
linker happy. Not the cleanest, but seems to get things going for now.